Move get-microservices-metdata to explore
