+++ cmd/tools/automate/root.go
@@ -14,7 +14,6 @@ Tools to slowly help make it easier to live in a manual and semi-automated world
 
 func init() {
 	RootCmd.AddCommand(pullDolittleConfigCMD)
-	RootCmd.AddCommand(getMicroservicesMetaDataCMD)
 	RootCmd.AddCommand(addPlatformConfigCMD)
 	RootCmd.AddCommand(importDolittleConfigMapsCMD)
 	RootCmd.AddCommand(pullMicroserviceDeploymentCMD)
