+++ cmd/tools/explore/root.go
@@ -15,4 +15,5 @@ Tools to Explore the cluster`,
 func init() {
 	RootCmd.AddCommand(ingressCMD)
 	RootCmd.AddCommand(dolittleResourcesCMD)
+	RootCmd.AddCommand(microservicesCMD)
 }
